requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
numpy==1.23.4
pandas==1.5.1
progress==1.6
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
Werkzeug==2.2.2
